Added Gyroscope to elevator, so it can be opened by either two fingers or shake
Added a threshold where if the elevator cannot be open it will break by itself when exceeding the threshold, and a break sound is added.
Ease of opening the elevator has been increased so it's hard to open (unless you swipe frantically or smash your phone to the floor) and made it easier to reach the threshold.